Renamed Tester project to Demo
Updated Readme.md code snippets to reference actual method names


nonfunctional refactoring of test cases

mostly renaming of classes to a somehwat standard naming scheme and making testcases sealed by default when not static
Removed examples from readme.md  that are not actually implemented


Updated snippets some more, to reflect currently implemented state
