+++ Decorator.Demo/Client.cs
@@ -1,10 +1,10 @@
 using Decorator.Attributes;
-using Decorator.Tester.MessageTypes;
+using Decorator.Demo.MessageTypes;
 
 using System;
 using System.Collections.Generic;
 
-namespace Decorator.Tester
+namespace Decorator.Demo
 {
 	public class Client
 	{
