+++ Decorator.Demo/GlobalSuppressions.cs
