+++ Decorator.Tests/Util/TestMessage.cs
