+++ Decorator.sln
@@ -1,15 +1,15 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2035
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Decorator.Tester", "Decorator.Tester\Decorator.Tester.csproj", "{AB74D9DB-3566-43D7-AA58-F2FE3EAA7850}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Decorator", "Decorator\Decorator.csproj", "{94C63DCE-C117-4EA7-86D9-AB2C1831AFED}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Decorator.Tests", "Decorator.Tests\Decorator.Tests.csproj", "{1D8FD7C7-0564-451E-8E0E-5B1FF03156FC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Decorator.Benchmarks", "Decorator.Benchmarks\Decorator.Benchmarks.csproj", "{7BBC2064-7985-4E00-8163-5B785B419DE4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Decorator.Demo", "Decorator.Demo\Decorator.Demo.csproj", "{C2B9A8D5-F5A9-4431-8A52-67589545EC7F}"
 EndProject
 Global
 	GlobalSection(Performance) = preSolution
@@ -20,10 +20,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AB74D9DB-3566-43D7-AA58-F2FE3EAA7850}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AB74D9DB-3566-43D7-AA58-F2FE3EAA7850}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AB74D9DB-3566-43D7-AA58-F2FE3EAA7850}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AB74D9DB-3566-43D7-AA58-F2FE3EAA7850}.Release|Any CPU.Build.0 = Release|Any CPU
 		{94C63DCE-C117-4EA7-86D9-AB2C1831AFED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{94C63DCE-C117-4EA7-86D9-AB2C1831AFED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94C63DCE-C117-4EA7-86D9-AB2C1831AFED}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -36,6 +32,10 @@ Global
 		{7BBC2064-7985-4E00-8163-5B785B419DE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7BBC2064-7985-4E00-8163-5B785B419DE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7BBC2064-7985-4E00-8163-5B785B419DE4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2B9A8D5-F5A9-4431-8A52-67589545EC7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2B9A8D5-F5A9-4431-8A52-67589545EC7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2B9A8D5-F5A9-4431-8A52-67589545EC7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2B9A8D5-F5A9-4431-8A52-67589545EC7F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
